api: sdn: show full device name of firewall bridge ports
The bridges endpoint also returns the firewall ports of the bridges.
From the firewall port interface we extract the vmid and the vm network
device index. Instead of just showing a number, return the whole
`net{index}` name. This makes it easier to understand as the vm
configuration also shows `net{index}` as the key of the interfaces.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Reported-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20251117140121.160192-1-g.goller@proxmox.com