fix #6806: vlan: qinq: fix bridge port detection
When creating a vlan zone and vnet, pve-network looks at all the
physical bridge ports (slaves) and adds them to the generated vlan
bridge. The zone plugin gets all the bridge interfaces using
`/sys/class/net` and then filters them using a regex. With the
introduction of network interface pinning, the restrictions on network
interface names have gotten more liberal - they're not required to
have specific prefixes anymore. The check for physical interfaces in
the zones plugin needs to be adjusted to reflect those changes,
otherwise the generated SDN configuration does not contain any pinned
physical ports and therefore doesn't work. Use the provided helper
from PVE::IPRoute2 instead, that adheres to the new naming policy and
uses `ip link` to determine the physical ports of the bridge, instead
of relying on a regex.

This improves the previous commit 4f19480b - which only allowed the
nic / if prefixes, which solved the issue when using the default
prefix, but not when using a custom prefix.

Fixes: 4f19480b04315afb5dc23e0130463acaea35db18
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251210184236.448600-5-s.hanreich@proxmox.com